Correção e adptação para a Nova API e tbm integração com o Novo Front

+++ back_end/novaalianca/src/main/java/com/condominio/novaalianca/repositories/CobrancaExtraRepository.java
@@ -10,5 +10,5 @@ import org.springframework.stereotype.Repository;
 public interface CobrancaExtraRepository extends JpaRepository<CobrancaExtra,Long> {
 
     @Query(value = "Select ce FROM CobrancaExtra ce where ce.unidade = :uni AND ce.mesReferencia = :mesReferencia ")
-    CobrancaExtra findByidUnidadeAndMesReferencia(Unidade uni, Integer mesReferencia);
+    CobrancaExtra findByidUnidadeAndMesReferencia(Unidade uni, Long mesReferencia);
 }
